+++ Domain/RDD.Domain/IEntityBase.cs
@@ -8,5 +8,5 @@
 
     public interface IEntityBase<TKey> : IEntityBase, IPrimaryKey<TKey> { }
 
-    public interface IEntityBase<TEntity, TKey> : IEntityBase<TKey>, ICloneable<TEntity> { }
+    public interface IEntityBase<TEntity, TKey> : IEntityBase<TKey>{ }
 }