+++ docs/CHANGELOG.md
@@ -16,6 +16,8 @@
  - **Modification**: `WebController` constructor now takes an `IQueryParser` and a `ICandidateParser` instead of an `ApiHelper`.
  - **Modification**: Multiple Put now returns a `ISelection` instead of enumerable, for consistency. Has potential impact on the front end.
  - **Removed**: Unused metadata.paging in returned json
+ - **Removed**: `IClonable<>` interface & Clone() method
+ - **Removed**: `Query<TEntity> query` parameter removed from prototype of ReadOnlyRepository Set() method 
  - **Modification**: ValidateEntity on RestCollection is now ValidateEntityAsync
  - **Modification**: RDD namespace renamed to Rdd
  - **Modification**: UseRDD() and AddRDD() extension methods renamed to UseRdd() and AddRdd()